requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
anthropic>=0.40.0
notion-client>=2.2.1
gitpython>=3.1.40
python-dotenv>=1.0.0
